Move bot pic_ methods into tweetpic.rb
Consolidates stuff.
